refactor: move user feedback related components to common directory
